backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile dev.in
#
aioredis==1.3.1
    # via channels-redis
asgiref==3.4.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.1
    # via aioredis
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
boto3==1.20.6
    # via -r base.in
botocore==1.23.6
    # via
    #   boto3
    #   s3transfer
cffi==1.15.0
    # via cryptography
channels==3.0.4
    # via
    #   -r base.in
    #   channels-redis
channels-redis==3.3.1
    # via -r base.in
constantly==15.1.0
    # via twisted
cryptography==35.0.0
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
django==3.2.9
    # via
    #   -r base.in
    #   channels
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cleanup==5.2.0
    # via -r base.in
django-simple-history==3.0.0
    # via -r base.in
django-storages==1.12.3
    # via -r base.in
djangorestframework==3.12.4
    # via
    #   -r base.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.0.0
    # via -r base.in
hiredis==2.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   hyperlink
    #   twisted
incremental==21.3.0
    # via twisted
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
msgpack==1.0.2
    # via channels-redis
psycopg2-binary==2.9.2
    # via -r base.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.21
    # via cffi
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pyopenssl==21.0.0
    # via twisted
python-dateutil==2.8.2
    # via botocore
pytz==2021.3
    # via django
s3transfer==0.5.0
    # via boto3
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   pyopenssl
    #   python-dateutil
    #   service-identity
sqlparse==0.4.2
    # via django
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==4.0.0
    # via
    #   async-timeout
    #   twisted
urllib3==1.26.7
    # via botocore
whitenoise==5.3.0
    # via -r base.in
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
